+++ cmake/options.cmake
@@ -90,15 +90,9 @@ add_feature_info("Trajectory plugin" BUILD_PLUGIN_TRAJECTORY
     "Compute trajectory from LAS input")
 
 option(BUILD_PLUGIN_E57
-        "Choose if e57 ui support should be built" FALSE)
+    "Choose if e57 ui support should be built" FALSE)
 add_feature_info("E57 plugin" BUILD_PLUGIN_E57
-        "read/write data to and from e57 format")
-
-option(BUILD_PLUGIN_STRAIGHTEN
-    "Choose if straighten support should be built" FALSE)
-add_feature_info("straighten plugin" BUILD_PLUGIN_STRAIGHTEN
-    "Straighten/unstraighten point cloud along polyline")
-
+    "read/write data to and from e57 format")
 
 option(BUILD_TOOLS_NITFWRAP "Choose if nitfwrap tool should be built" FALSE)
