+++ filters/StraightenFilter.hpp
