+++ filters/private/straighten/Polyline.cpp
