+++ filters/private/straighten/Polyline.hpp
