+++ filters/private/straighten/Utils.cpp
