+++ filters/private/straighten/Utils.hpp
