+++ plugins/CMakeLists.txt
@@ -74,7 +74,3 @@ endif()
 if(BUILD_PLUGIN_E57)
     add_subdirectory(e57)
 endif()
-
-if(BUILD_PLUGIN_STRAIGHTEN)
-    add_subdirectory(straighten)
-endif()
